pyproject.toml
[project]
name = "dde-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyterlab (>=4.4.7,<5.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)",
    "wget (>=3.2,<4.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)"
]

[tool.poetry]
packages = [{include = "dde_project", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
